Finished XML commenting; Completed testing of current endpoints.
Note, unable to check most of the User endpoints, as this requires special access on the Client-ID and a regular ID will not work. Also have currently disabled checks for EmoteSetRequest (to retrieve an emote set by its ID) as this request appears to be not responding - producing Internal Server Error.